ktp_dummy_interface/pip_requirement.txt
msgpack-python>=0.4.8
paho-mqtt>=2.0